Use integer quantities in ship summary contract

+++ backend/IQBF.Application/DTOs/Dashboard/ShipSummaryDto.cs
@@ -3,11 +3,11 @@ namespace IQBF.Application.DTOs.Dashboard;
 public record ShipSummaryDto(
     Guid ShipId,
     string ShipName,
-    decimal TotalQuantity,
-    decimal ReceivedQuantity,
-    decimal DispatchedQuantity,
-    decimal AvailableQuantity,
-    decimal PendingReception,
+    int TotalQuantity,
+    int ReceivedQuantity,
+    int DispatchedQuantity,
+    int AvailableQuantity,
+    int PendingReception,
     decimal ReceptionProgress,
     decimal DispatchProgress,
     IReadOnlyCollection<BLBalanceDto> BLs
